fix: MyPy型エラー142件を完全解決 - 型安全性の大幅改善
## 変更内容

### MyPy型チェックの完全対応
- 142件の型エラーを0件に削減
- pre-pushフックでMyPyチェックを再有効化
- 全モジュールで厳密な型注釈を実装

### SQLAlchemy 2.0対応
- declarative_base()から新しいDeclarativeBaseへ移行
- 型ヒントの改善とモダンな実装パターンの採用

### Pydantic v2互換性修正
- 非推奨の'env'パラメータを削除
- Config.env_fileを直接指定する新しいパターンに更新
- カスタムバリデータの型シグネチャを修正（str | bool返り値対応）

### 型安全性の改善
- JWT認証ハンドラーの到達不能コード修正
- WebSocket接続マネージャーの型注釈強化
- スキーマバリデータの型エラー解決
- リポジトリレイヤーの型注釈改善

### 影響範囲
- API認証・ルーティング層
- データインポート・パース処理
- 特徴量抽出モジュール
- データ処理パイプライン
- WebSocketリアルタイム通信

## テスト結果
- MyPy: エラー0件（142件から改善）
- Flake8: 全件パス
- 既存テストの互換性維持

+++ src/features/extractors/jockey_trainer.py
@@ -451,7 +451,7 @@ class JockeyTrainerFeatureExtractor:
             else:
                 scores.append(0.1)
 
-        return np.mean(scores)
+        return float(np.mean(scores))
 
     def _get_recent_data(self, data: pd.DataFrame, days: int = 30) -> pd.DataFrame:
         """最近のデータを取得
